refactor: split sidebar in a dedicated component; make api request
