+++ src/routes/_student/form/index.tsx
@@ -1,72 +1,37 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { Tabs, TabsContent } from "@/components/ui/tabs";
-import { GalleryVerticalEnd } from "lucide-react";
-import { useState, type ChangeEvent, useMemo } from "react";
+import { useState, useCallback } from "react";
 import { Checkbox } from "@/components/ui/checkbox";
-import { Label } from "@/components/ui/label";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { Button } from "@/components/ui/button";
-import formData, { type FormQuestion, type FormOption } from "./-data";
+import formData, { type FormQuestion } from "./-data";
 import { formSchema, getInitialValues, type FormValues } from "./-schema";
-import { useForm, useWatch, type FieldErrors } from "react-hook-form";
+import { useForm, type FieldErrors } from "react-hook-form";
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { toast } from "sonner";
+import { useMutation } from "@tanstack/react-query";
+import { createStudentRegistrationMutationOptions } from "@/mutations/create-student-registration";
+import { CreateStudentRegistrationFormSidebar } from "./-sidebar";
 
 export const Route = createFileRoute("/_student/form/")({
   component: StudentRegistrationForm,
 });
 
 export function StudentRegistrationForm() {
+  const { mutateAsync } = useMutation(createStudentRegistrationMutationOptions);
   const form = useForm({
     resolver: zodResolver(formSchema),
     defaultValues: getInitialValues(),
-  })
-  const formValues = useWatch({ control: form.control });
+
+  })  
+ 
   const [activeTab, setActiveTab] = useState(formData.sections[0].id);
 
-  const isQuestionAnswered = (question: FormQuestion): boolean => {
-    const value = formValues[question.id as keyof FormValues];
-    
-    if (!question.required) return true;
-    
-    switch (question.type) {
-      case "text":
-      case "email":
-      case "textarea":
-        return typeof value === "string" && value.trim().length > 0;
-      case "radio":
-        return typeof value === "string" && value.length > 0;
-      case "checkbox":
-        return Array.isArray(value) && value.length > 0;
-      case "checkbox-single":
-        return value === true;
-      case "file":
-        return value instanceof File;
-      default:
-        return false;
-    }
-  };
-
-  const sectionProgress = useMemo(() => {
-    return formData.sections.map(section => {
-      const requiredQuestions = section.questions.filter(q => q.required);
-      const totalRequired = requiredQuestions.length;
-      const answeredRequired = requiredQuestions.filter(q => isQuestionAnswered(q)).length;
-      
-      return {
-        sectionId: section.id,
-        answered: answeredRequired,
-        total: totalRequired,
-        percentage: totalRequired > 0 ? Math.round((answeredRequired / totalRequired) * 100) : 100
-      };
-    });
-  }, [formValues]);
-
-  function renderQuestion(question: FormQuestion) {
-    switch (question.type) {
+  const renderQuestion = useCallback((question: FormQuestion) => {
+      switch (question.type) {
       case "text":
       case "email":
         return (
@@ -107,6 +72,7 @@ export function StudentRegistrationForm() {
             <FormField
               control={form.control}
               name={question.id}
+              key={question.id}
               render={({ field }) => (
                 <FormItem className="space-y-3">
                   <FormLabel>{question.question}</FormLabel>
@@ -126,7 +92,6 @@ export function StudentRegistrationForm() {
                             </FormLabel>
                           </FormItem>
                         ))}
-                      
                       </RadioGroup>
                     </FormControl>
                     <FormMessage />
@@ -138,6 +103,7 @@ export function StudentRegistrationForm() {
         return (
           <FormField
             control={form.control}
+            key={question.id}
             name={question.id}
             render={({ field }) => {
               const valueSet = new Set(field.value || []);
@@ -163,7 +129,7 @@ export function StudentRegistrationForm() {
                             />
                           </FormControl>
                           <FormLabel className="font-normal">
-                            {option.label} {option.id}
+                            {option.label}
                           </FormLabel>
                         </FormItem>
                       ))}
@@ -179,6 +145,7 @@ export function StudentRegistrationForm() {
         return (
           <FormField
             control={form.control}
+            key={question.id}
             name={question.id}
             render={({ field }) => (
               <FormItem className="space-y-3">
@@ -205,6 +172,7 @@ export function StudentRegistrationForm() {
         return (
           <FormField
             control={form.control}
+            key={question.id}
             name={question.id}
             render={({ field }) => (
               <FormItem className="space-y-3">
@@ -227,10 +195,10 @@ export function StudentRegistrationForm() {
       default:
         return null;
     }
-  }
+  }, [])
 
-  const onSubmit = (values: FormValues) => {
-    console.log("Form submitted:", values);
+  const onSubmit = async (values: FormValues) => {
+    await mutateAsync({ editalId: 1, data: {answer: values} });
   };
 
   const onError = (errors: FieldErrors<FormValues>) => {
@@ -240,62 +208,7 @@ export function StudentRegistrationForm() {
   return (
     <div className="flex h-full">
       {/* SIDEBAR */}
-      <div className="flex flex-col h-full w-64 border-r">
-        <div className="flex items-center gap-2 p-4">
-          <div className="bg-primary rounded-lg p-2">
-            <GalleryVerticalEnd className="text-secondary size-4" />
-          </div>
-          <p className="text-xs font-semibold">
-            Cadastramento
-            <br />
-            Socioeconômico - 2025.1
-          </p>
-        </div>
-
-        <div className="flex flex-col gap-2 pr-2">
-          {formData.sections.map((section, idx) => {
-            const progress = sectionProgress.find(p => p.sectionId === section.id);
-            const isComplete = progress?.percentage === 100;
-            
-            return (
-              <button
-                key={section.id}
-                onClick={() => setActiveTab(section.id)}
-                className={`flex flex-col gap-2 p-2 pl-3 min-h-12 border rounded-tr-md rounded-br-md transition 
-                  ${
-                    activeTab === section.id
-                      ? "bg-blue-100"
-                      : "bg-card hover:bg-slate-100"
-                  }`}
-              >
-                <div className="flex items-center justify-between w-full">
-                  <p className="text-xs text-left max-w-40">{section.title}</p>
-                  <div className={`flex items-center justify-center size-6 rounded-full ${
-                    isComplete ? "bg-green-500 text-white" : "bg-slate-300"
-                  }`}>
-                    <p className="text-xs font-bold">{idx + 1}</p>
-                  </div>
-                </div>
-                
-                {/* Barra de progresso */}
-                <div className="w-full bg-gray-200 rounded-full h-1.5">
-                  <div
-                    className={`h-1.5 rounded-full transition-all duration-300 ${
-                      isComplete ? "bg-green-500" : "bg-blue-500"
-                    }`}
-                    style={{ width: `${progress?.percentage || 0}%` }}
-                  />
-                </div>
-                
-                {/* Texto do progresso */}
-                <p className="text-[10px] text-gray-500 text-left">
-                  {progress?.answered}/{progress?.total} respondidas
-                </p>
-              </button>
-            );
-          })}
-        </div>
-      </div>
+      <CreateStudentRegistrationFormSidebar activeTab={activeTab} form={form} changeTab={setActiveTab} />
 
       {/* TABS */}
       <Form {...form}>
