+++ src/types/create-student-registration-dto.ts
@@ -1,0 +1,3 @@
+export type CreateStudentRegistrationDTO = {
+  answer: Record<string, any>;
+};
